requirements.txt
asyncio
uvicorn
fastapi
bcrypt
asyncpg
pydantic
python-dotenv==0.19.1
